+++ packages/headless/doc-parser/doc-parser.ts
@@ -11,6 +11,17 @@ const entryPoint = apiPackage.entryPoints[0];
 
 const controllers: ControllerConfiguration[] = [
   {
+    initializer: 'buildFacet',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/facet/facet.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/facet/facet.fn.tsx',
+      ],
+    },
+  },
+  {
     initializer: 'buildPager',
     samplePaths: {
       react_class: [
@@ -29,6 +40,10 @@ const controllers: ControllerConfiguration[] = [
       ],
     },
   },
+  // {
+  //   initializer: 'buildCategoryFacet',
+  //   samplePaths: {},
+  // },
   {
     initializer: 'buildDidYouMean',
     samplePaths: {
@@ -85,6 +100,17 @@ const controllers: ControllerConfiguration[] = [
     },
   },
   {
+    initializer: 'buildTab',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/tab/tab.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/tab/tab.fn.tsx',
+      ],
+    },
+  },
+  {
     initializer: 'buildSearchStatus',
     samplePaths: {
       react_class: [
@@ -92,6 +118,48 @@ const controllers: ControllerConfiguration[] = [
       ],
       react_fn: [
         'packages/samples/headless-react/src/components/search-status/search-status.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildSearchBox',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/search-box/search-box.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/search-box/search-box.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildInteractiveResult',
+    samplePaths: {
+      react_class: [],
+      react_fn: [
+        'packages/samples/headless-react/src/components/result-list/result-link.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildResultsPerPage',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/results-per-page/results-per-page.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/results-per-page/results-per-page.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildSearchParameterManager',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/search-parameter-manager/search-parameter-manager.ts',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/search-parameter-manager/search-parameter-manager.ts',
       ],
     },
   },