+++ packages/headless/doc-parser/src/interface-resolver.ts
@@ -62,7 +62,7 @@ function resolveInheritedMembers(
 ) {
   return apiInterface.extendsTypes
     .map((m) => {
-      const typeName = extractTypeName(m.excerpt);
+      const typeName = extractSearchableTypeName(m.excerpt);
       const inheritedInterface = findApi(entry, typeName) as ApiInterface;
       return resolveInterfaceMembers(entry, inheritedInterface);
     })
@@ -122,7 +122,8 @@ function buildObjEntityFromProperty(
   p: ApiPropertySignature
 ) {
   const typeName = extractTypeName(p.propertyTypeExcerpt);
-  const apiInterface = findApi(entry, typeName) as ApiInterface;
+  const searchableTypeName = extractSearchableTypeName(p.propertyTypeExcerpt);
+  const apiInterface = findApi(entry, searchableTypeName) as ApiInterface;
   const members = resolveInterfaceMembers(entry, apiInterface);
   const entity = buildEntityFromProperty(p);
 
@@ -134,8 +135,8 @@ function buildEntityFromPropertyAndResolveTypeAlias(
   p: ApiPropertySignature
 ): Entity {
   const entity = buildEntityFromProperty(p);
-  const alias = extractTypeName(p.propertyTypeExcerpt);
-  const typeAlias = findApi(entry, alias) as ApiTypeAlias;
+  const searchableTypeName = extractSearchableTypeName(p.propertyTypeExcerpt);
+  const typeAlias = findApi(entry, searchableTypeName) as ApiTypeAlias;
   const type = typeAlias.typeExcerpt.text;
 
   return {...entity, type};
@@ -176,8 +177,8 @@ function buildEntityFromParamAndResolveTypeAlias(
   p: Parameter
 ): Entity {
   const entity = buildParamEntity(p);
-  const alias = extractTypeName(p.parameterTypeExcerpt);
-  const typeAlias = findApi(entry, alias) as ApiTypeAlias;
+  const searchableTypeName = extractSearchableTypeName(p.parameterTypeExcerpt);
+  const typeAlias = findApi(entry, searchableTypeName) as ApiTypeAlias;
   const type = typeAlias.typeExcerpt.text;
 
   return {...entity, type};
@@ -185,7 +186,8 @@ function buildEntityFromParamAndResolveTypeAlias(
 
 function buildObjEntityFromParam(entryPoint: ApiEntryPoint, p: Parameter) {
   const typeName = extractTypeName(p.parameterTypeExcerpt);
-  const apiInterface = findApi(entryPoint, typeName) as ApiInterface;
+  const searchableTypeName = extractSearchableTypeName(p.parameterTypeExcerpt);
+  const apiInterface = findApi(entryPoint, searchableTypeName) as ApiInterface;
   const members = resolveInterfaceMembers(entryPoint, apiInterface);
   const entity = buildParamEntity(p);
 
@@ -193,5 +195,9 @@ function buildObjEntityFromParam(entryPoint: ApiEntryPoint, p: Parameter) {
 }
 
 function extractTypeName(excerpt: Excerpt) {
+  return excerpt.text.replace(/\[\]/, '');
+}
+
+function extractSearchableTypeName(excerpt: Excerpt) {
   return excerpt.spannedTokens[0].text;
 }