+++ packages/headless/src/controllers/results-per-page/headless-results-per-page.ts
@@ -11,7 +11,7 @@ import {
   PaginationSection,
 } from '../../state/state-sections';
 import {validateInitialState} from '../../utils/validate-payload';
-import {buildController} from '../controller/headless-controller';
+import {buildController, Controller} from '../controller/headless-controller';
 
 const initialStateSchema = new Schema({
   numberOfResults: new NumberValue({min: 0}),
@@ -21,27 +21,61 @@ export interface ResultsPerPageProps {
   /**
    * The initial state that should be applied to this `ResultsPerPage` controller.
    */
-  initialState: Partial<ResultsPerPageInitialState>;
+  initialState?: ResultsPerPageInitialState;
 }
 
 export interface ResultsPerPageInitialState {
   /**
    * The initial number of results to register in state.
    */
-  numberOfResults: number;
+  numberOfResults?: number;
 }
 
 /**
  * The `ResultsPerPage` controller allows the end user to choose how many results to display per page.
  */
-export type ResultsPerPage = ReturnType<typeof buildResultsPerPage>;
-/** A scoped and simplified part of the headless state that is relevant to the `ResultsPerPage` controller. */
-export type ResultsPerPageState = ResultsPerPage['state'];
+export interface ResultsPerPage extends Controller {
+  /**
+   * Updates the number of results to request per page.
+   *
+   * @param num - The number of results.
+   */
+  set(num: number): void;
 
+  /** Checks whether the number of results per page is equal to the specified number.
+   *
+   * @param num - The number of results.
+   * @returns `true` if the number of results is equal to the passed value, and `false` otherwise.
+   */
+  isSetTo(num: number): boolean;
+
+  /**
+   * The state of the `ResultsPerPage` controller.
+   */
+  state: ResultsPerPageState;
+}
+
+/**
+ * A scoped and simplified part of the headless state that is relevant to the `ResultsPerPage` controller.
+ * */
+export interface ResultsPerPageState {
+  /**
+   * The number of results per page.
+   * */
+  numberOfResults: number;
+}
+
+/**
+ * Creates a `ResultsPerPage` controller instance.
+ *
+ * @param engine - The headless engine.
+ * @param props - The configurable `ResultsPerPage` properties.
+ * @returns A `ResultsPerPage` controller instance.
+ */
 export function buildResultsPerPage(
   engine: Engine<PaginationSection & ConfigurationSection>,
-  props: Partial<ResultsPerPageProps> = {}
-) {
+  props: ResultsPerPageProps = {}
+): ResultsPerPage {
   const controller = buildController(engine);
   const {dispatch} = engine;
 
@@ -61,28 +95,17 @@ export function buildResultsPerPage(
   return {
     ...controller,
 
-    /**
-     * The state of the `ResultsPerPage` controller.
-     */
     get state() {
       return {
-        /** The number of results per page. */
         numberOfResults: engine.state.pagination.numberOfResults,
       };
     },
 
-    /** Updates the number of results to request per page.
-     * @param num The number of results.
-     */
     set(num: number) {
       dispatch(updateNumberOfResults(num));
       dispatch(executeSearch(logPagerResize()));
     },
 
-    /** Checks whether the number of results per page is equal to the specified number.
-     * @param num The number of results.
-     * @returns `true` if the number of results is equal to the passed value, and `false` otherwise.
-     */
     isSetTo(num: number) {
       return num === this.state.numberOfResults;
     },