+++ CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0-alpha.27](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.26...v0.7.0-alpha.27) (2021-03-02)
+
+
+### Features
+
+* **atomic:** complete atomic-result-printable-uri component ([#566](https://github.com/coveo/ui-kit/issues/566)) ([f6ea7d2](https://github.com/coveo/ui-kit/commit/f6ea7d2275717e8b743cfe01a1f6d6528061f6ee))
+
+
+
+
+
+# [0.7.0-alpha.26](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.25...v0.7.0-alpha.26) (2021-03-02)
+
+**Note:** Version bump only for package ui-kit
+
+
+
+
+
+# [0.7.0-alpha.25](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.24...v0.7.0-alpha.25) (2021-03-01)
+
+
+### Features
+
+* **atomic:** make default CSS variables available at the root ([#551](https://github.com/coveo/ui-kit/issues/551)) ([950c1fc](https://github.com/coveo/ui-kit/commit/950c1fc58cbbcebc06265fc66d6c79c39158271e))
+* **atomic:** remove atomic-result-excerpt component from DOM when empty ([#550](https://github.com/coveo/ui-kit/issues/550)) ([28baaae](https://github.com/coveo/ui-kit/commit/28baaae439c9dbf5d7df139013615b57c2d99733))
+
+
+
+
+
 # [0.7.0-alpha.24](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.23...v0.7.0-alpha.24) (2021-02-26)